requirements.txt
keras==3.7.0
numpy==1.26.4
pillow==11.0.0
protobuf==4.25.5
streamlit==1.41.1
tensorflow==2.16.1
opencv-python==4.10.0.84